Fixes to get the display working in colour
It's alive! Alive!  Enough fixes to get it working in 262k colours.
It's a little laggy, so maybe I should go back down to 65k colours, but
we'll see.  It's good enough.

+++ ssd1351.hpp
@@ -53,7 +53,7 @@ class Ssd1351 : public Display {
         Ssd1351(const char *spi_dev, int cs, int dc, int rst = -1);
         void drawImage(libcamera::Span<uint8_t>& data) override;
         void drawPixel(int16_t x, int16_t y, uint16_t color) override;
-        void fillWithColour(uint16_t colour) override;
+        void fillWithColour(uint32_t colour) override;
         void displayOff() override;
     protected:
         void setAddrWindow(uint16_t x1, uint16_t y1, uint16_t w, uint16_t h) override;
