+++ tp28017.hpp
@@ -70,7 +70,7 @@ class Tp28017 : public Display {
         Tp28017(const char *spi_dev, int cs, int dc, int rst = -1);
         void drawImage(libcamera::Span<uint8_t>& data) override;
         void drawPixel(int16_t x, int16_t y, uint16_t color) override;
-        void fillWithColour(uint16_t colour) override;
+        void fillWithColour(uint32_t colour) override;
         void displayOff() override;
     protected:
         void setAddrWindow(uint16_t x1, uint16_t y1, uint16_t w, uint16_t h) override;
